fix: move init command to separate package
