+++ packages/webpack-cli/lib/utils/resolve-command.js
@@ -1,13 +1,15 @@
 const { yellow, cyan } = require('colorette');
-const logger = require('../utils/logger');
-const { packageExists } = require('../utils/package-exists');
-const { promptInstallation } = require('../utils/prompt-installation');
+const logger = require('./logger');
+const { packageExists } = require('./package-exists');
+const { promptInstallation } = require('./prompt-installation');
 
 const packagePrefix = '@webpack-cli';
 
 const run = async (name, ...args) => {
     const scopeName = packagePrefix + '/' + name;
+
     let pkgLoc = packageExists(scopeName);
+
     if (!pkgLoc) {
         try {
             pkgLoc = await promptInstallation(`${scopeName}`, () => {
@@ -17,7 +19,18 @@ const run = async (name, ...args) => {
             logger.error(`Action Interrupted, use ${cyan('webpack-cli help')} to see possible commands.`);
         }
     }
-    return pkgLoc ? require(scopeName).default(...args) : null;
+
+    if (!pkgLoc) {
+        return;
+    }
+
+    let mod = require(scopeName);
+
+    if (mod.default) {
+        mod = mod.default;
+    }
+
+    return mod(...args);
 };
 
 module.exports = run;
